LayoutTests/svg/custom/selectSubString-expected.txt

PASS svgText.selectSubString(-1, 0) threw exception IndexSizeError (DOM Exception 1): The index is not in the allowed range..
PASS svgText.getSubStringLength(0, -1) is 60
PASS svgText.getSubStringLength(3, 0) threw exception IndexSizeError (DOM Exception 1): The index is not in the allowed range..
PASS successfullyParsed is true

TEST COMPLETE
